// client/src/types.ts
export interface GuildMembership {
  guild_id: number;
  member_id: number;
  permissions: permission_type[];
  status: registration_status_type;
  created_at: string;
  updated_at: string;
}
export interface QuestMembership {
  quest_id: number;
  member_id: number;
  permissions: permission_type[];
  confirmed: boolean;
  created_at: string;
  updated_at: string;
}
export interface Casting {
  guild_id: number;
  quest_id: number;
  member_id: number;
  permissions: permission_type[];
  status: registration_status_type;
  created_at: string;
  updated_at: string;
}
// ...
export interface BaseConversationNode {
  node_type: ibis_node_type_type;
  status: publication_state_type;
  created_at: string;
  published_at: string;
  updated_at: string;
  title: string;
  description: string;
  url: string;
  meta: meta_state_type;
  draft_for_role_id: number;
}
// ...
export type generic_id = number | string;
export interface PseudoNode extends BaseConversationNode {
  id: generic_id;
  quest_id: generic_id;
  guild_id?: generic_id;
  creator_id: generic_id;
  parent_id?: generic_id;
  children?: PseudoNode[];
}
// ...
export interface GuildMemberAvailableRole {
  guild_id: number;
  member_id: number;
  role_id: number;
}

export interface CastingRole {
  guild_id: number;
  member_id: number;
  role_id: number;
  quest_id: number;
}